feat: replace bug-catch with plausible
