+++ app/env.sample.js
@@ -10,6 +10,9 @@ export default {
 	ANIMATION_LONG: 3000,
 	ANIMATION_SHORT: 1000,
 	BUGCATCH_ENABLE: false,
+	PLAUSIBLE_ENABLE: false,
+	PLAUSIBLE_HOST: 'app-ambient-tv.example.com',
+	PLAUSIBLE_API_URL: 'https://plausible.io',
 	APP_SERVER_URL: '',
 	FETCH_METHOD: '',
 	IMAGE_TIMER: 120, //  1 = 1s, 120 = 2m
