+++ app/package.json
@@ -1,6 +1,6 @@
 {
     "name": "ambient-tv",
-    "version": "1.4.21",
+    "version": "1.5.21",
     "main": "node_modules/expo/AppEntry.js",
     "scripts": {
         "start": "expo start --tunnel",
@@ -16,13 +16,13 @@
         "web:build": "expo export --platform web"
     },
     "dependencies": {
-        "@bug-catch/react-native": "^2.2.17",
         "@react-native-async-storage/async-storage": "1.18.2",
         "axios": "^1.4.0",
         "chance": "^1.1.11",
         "expo": "~49.0.5",
         "expo-av": "~13.4.1",
         "expo-build-properties": "~0.8.3",
+        "expo-constants": "~14.4.2",
         "expo-font": "~11.4.0",
         "expo-keep-awake": "~12.3.0",
         "expo-linear-gradient": "~12.3.0",
